Downgrade to a stable version of Tauri, Configure Rust to be a binary instead of a library build
