+++ package.json
@@ -10,23 +10,18 @@
     "tauri": "tauri"
   },
   "dependencies": {
-    "@tauri-apps/api": "^2.0.0-alpha.4",
-    "@tauri-apps/plugin-shell": "^2.0.0-alpha.0",
-    "@tauri-apps/plugin-window": "^2.0.0-alpha.0",
+    "@tauri-apps/api": "^1.4.0",
     "axios": "^1.5.0",
-    "global": "^4.4.0",
-    "pnpm": "^8.7.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.15.0",
     "styled-components": "^6.0.7"
   },
   "devDependencies": {
-    "@tauri-apps/cli": "^2.0.0-alpha.9",
+    "@tauri-apps/cli": "^1.4.0",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@vitejs/plugin-react": "^4.0.3",
-    "internal-ip": "^7.0.0",
     "typescript": "^5.0.2",
     "vite": "^4.4.4"
   }
