+++ src-tauri/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ARK-Rate-Desktop"
+name = "Ark-Rate-Desktop"
 version = "0.0.0"
 description = "A Tauri App"
 authors = ["you"]
@@ -7,19 +7,13 @@ license = ""
 repository = ""
 edition = "2021"
 
-[lib]
-name = "ARK_Rate_Desktop_lib"
-crate-type = ["staticlib", "cdylib", "rlib"]
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "2.0.0-alpha", features = [] }
+tauri-build = { version = "1.4", features = [] }
 
 [dependencies]
-tauri = { version = "2.0.0-alpha", features = [] }
-tauri-plugin-window = "2.0.0-alpha"
-tauri-plugin-shell = "2.0.0-alpha"
+tauri = { version = "1.4", features = ["shell-open"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0"
