+++ src-tauri/tauri.conf.json
@@ -1,19 +1,26 @@
 {
   "build": {
-    "beforeDevCommand": "pnpm dev",
-    "beforeBuildCommand": "pnpm build",
+    "beforeDevCommand": "yarn dev",
+    "beforeBuildCommand": "yarn build",
     "devPath": "http://localhost:1420",
     "distDir": "../dist",
     "withGlobalTauri": false
   },
   "package": {
-    "productName": "ARK-Rate-Desktop",
+    "productName": "Ark-Rate-Desktop",
     "version": "0.0.0"
   },
   "tauri": {
+    "allowlist": {
+      "all": false,
+      "shell": {
+        "all": false,
+        "open": true
+      }
+    },
     "bundle": {
       "active": true,
-      "targets": ["msi", "app", "dmg"],
+      "targets": "all",
       "identifier": "com.arkrate.desktop",
       "icon": [
         "icons/32x32.png",
@@ -30,15 +37,10 @@
       {
         "fullscreen": false,
         "resizable": true,
-        "title": "ARK-Rate-Desktop",
+        "title": "Ark-Rate-Desktop",
         "width": 800,
         "height": 600
       }
     ]
-  },
-  "plugins": {
-    "shell": {
-      "open": true
-    }
   }
 }
