+++ vite.config.ts
@@ -1,9 +1,5 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-import { internalIpV4 } from "internal-ip";
-
-// @ts-expect-error process is a nodejs global
-const mobile = !!/android|ios/.exec(process.env.TAURI_PLATFORM);
 
 // https://vitejs.dev/config/
 export default defineConfig(async () => ({
@@ -15,15 +11,7 @@ export default defineConfig(async () => ({
   clearScreen: false,
   // 2. tauri expects a fixed port, fail if that port is not available
   server: {
-    host: mobile ? "0.0.0.0" : false,
     port: 1420,
-    hmr: mobile
-      ? {
-          protocol: "ws",
-          host: await internalIpV4(),
-          port: 1421,
-        }
-      : undefined,
     strictPort: true,
   },
   // 3. to make use of `TAURI_DEBUG` and other env variables
